Still don't know why trying to read from the hashmap of material names to Material objects crashes

+++ Colour.cpp
@@ -7,24 +7,24 @@ Colour::Colour()
 	b = 0;
 }
 
-Colour::Colour(double nr, double ng, double nb)
+Colour::Colour(float nr, float ng, float nb)
 {
 	r = nr;
 	g = ng;
 	b = nb;
 }
 
-double Colour::getR()
+float Colour::getR()
 {
 	return r/255; // OpenGL r
 }
 
-double Colour::getG()
+float Colour::getG()
 {
 	return g/255; // OpenGL g
 }
 
-double Colour::getB()
+float Colour::getB()
 {
 	return b/255; // OpenGL b
 }
