+++ Colour.hpp
@@ -4,17 +4,17 @@
 class Colour
 {
 	private:
-		double r;
-		double g;
-		double b;
+		float r;
+		float g;
+		float b;
 
 	public:
 		Colour();
-		Colour(double nr, double ng, double nb);
+		Colour(float nr, float ng, float nb);
 		Colour(Colour& other); // Copy constructor
-		double getR();
-		double getG();
-		double getB();
+		float getR();
+		float getG();
+		float getB();
 };
 
 #endif
